chore: rename client-go to go-client
