+++ .bumpversion.go.cfg
@@ -19,7 +19,7 @@ replace = {new_version}
 search = {current_version}
 replace = {new_version}
 
-[bumpversion:file:experimental/client-go/go.mod]
+[bumpversion:file:experimental/go-client/go.mod]
 search = {current_version}
 replace = {new_version}
