+++ experimental/go-client/bindings/c/examples/train.c
