+++ experimental/go-client/bindings/core/embed-core.bin
