+++ experimental/go-client/bindings/core/generate.go
