+++ experimental/go-client/bindings/core/partialdata.go
