+++ experimental/go-client/bindings/cpp/examples/train.cpp
