+++ experimental/go-client/bindings/cpp/examples/train_session.cpp
