+++ experimental/go-client/bindings/scripts/format.sh
