+++ experimental/go-client/cmd/example/embed-core.bin
