+++ experimental/go-client/internal/gowandb/internal_runopts/runopts.go
@@ -2,7 +2,7 @@ package internal_runopts
 
 import (
 	"github.com/wandb/wandb/core/pkg/service"
-	"github.com/wandb/wandb/experimental/client-go/pkg/opts/runopts"
+	"github.com/wandb/wandb/experimental/go-client/pkg/opts/runopts"
 )
 
 func WithTelemetry(telemetry *service.TelemetryRecord) runopts.RunOption {
