+++ experimental/go-client/pkg/gowandb/run.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/wandb/wandb/core/pkg/service"
 	"github.com/wandb/wandb/core/pkg/utils"
-	"github.com/wandb/wandb/experimental/client-go/pkg/opts/runopts"
-	"github.com/wandb/wandb/experimental/client-go/pkg/runconfig"
+	"github.com/wandb/wandb/experimental/go-client/pkg/opts/runopts"
+	"github.com/wandb/wandb/experimental/go-client/pkg/runconfig"
 )
 
 type Settings map[string]interface{}
