+++ experimental/go-client/pkg/gowandb/session.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/wandb/wandb/experimental/client-go/internal/execbin"
-	"github.com/wandb/wandb/experimental/client-go/internal/launcher"
-	"github.com/wandb/wandb/experimental/client-go/pkg/opts/runopts"
-	"github.com/wandb/wandb/experimental/client-go/pkg/opts/sessionopts"
-	"github.com/wandb/wandb/experimental/client-go/pkg/settings"
+	"github.com/wandb/wandb/experimental/go-client/internal/execbin"
+	"github.com/wandb/wandb/experimental/go-client/internal/launcher"
+	"github.com/wandb/wandb/experimental/go-client/pkg/opts/runopts"
+	"github.com/wandb/wandb/experimental/go-client/pkg/opts/sessionopts"
+	"github.com/wandb/wandb/experimental/go-client/pkg/settings"
 )
 
 type Session struct {
